+++ primint/src/num.rs
@@ -1,5 +1,7 @@
 //! Mirrors the [`core::num`] module.
 
+mod nonmax;
 mod nonzero;
 
+pub use nonmax::NonMax;
 pub use nonzero::NonZero;