+++ primint/src/num/nonmax.rs
@@ -1,0 +1,152 @@
+use crate::num::NonZero;
+use crate::UnsignedPrimInt;
+
+/// An integer which is known not to be equal to the maximum value for the type.
+///
+/// Mirrors the [`core::num::NonZero`] type, but with a different forbidden value.
+/// This offers an alternative to the types in the [`nonmax`](https://docs.rs/nonmax) crate,
+/// which don't offer the nice generic syntax that [`core::num::NonZero`] does.
+///
+/// Currently, this type is limited to unsigned primitive integers.
+/// This restriction may be lifted in a future version.
+///
+/// # Safety
+/// As `T` is limited to builtin primitive integers,
+/// The correctness of this type can be relied upon for unsafe code.
+///
+/// It is guaranteed that `Option<NonZero<T>>` is the same size as `T`.
+/// However, the underlying representation of the type is not currently guaranteed.
+/// Right now, it is the numeric value plus one (so `NonMax::new(0)` is represented as 1).
+/// However, this may change in the future without notice (so even in patch versions).
+///
+/// Note that [`NonMax`] does not currently implement [`bytemuck::Contiguous`],
+/// as that would expose the underlying representation.
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
+#[repr(transparent)]
+pub struct NonMax<T: UnsignedPrimInt> {
+    value_plus_one: NonZero<T>,
+}
+impl<T: UnsignedPrimInt> NonMax<T> {
+    /// Creates a [`NonMax`] if the value is not zero,
+    /// or `None` if the value is zero.
+    #[inline]
+    pub fn new(val: T) -> Option<Self> {
+        if val == T::MAX {
+            None
+        } else {
+            // SAFETY: Just verified we are not the maximum value
+            Some(unsafe { Self::new_unchecked(val) })
+        }
+    }
+    /// Creates [`NonMax`] without checking whether the value is in bounds.
+    /// Undefined behavior occurs if the value is actually the maximum value for the type
+    ///
+    /// # Safety
+    /// Value must be less than the maximum value of `T`.
+    #[inline]
+    pub unsafe fn new_unchecked(val: T) -> Self {
+        // encountering this condition is UB, so a panic is a valid outcome
+        debug_assert!(val < T::MAX);
+        // Using unchecked_add might be slightly faster, but would require an increased MSRV or rust version macros
+        // Better to just wait for NonMax to be added to stdlib
+        // SAFETY: Adding one to a non-max value ends up with something nonzero
+        unsafe {
+            NonMax {
+                value_plus_one: NonZero::new_unchecked(T::wrapping_add(val, T::ONE)),
+            }
+        }
+    }
+
+    /// Get the underlying integer value.
+    ///
+    /// The result will never be the maximum value of `T`.
+    #[inline]
+    pub fn get(self) -> T {
+        // use of unchecked_sub here might be faster, but see above for comments
+        T::wrapping_sub(self.value_plus_one.get(), T::ONE)
+    }
+
+    /// The minimum value for this type.
+    ///
+    /// This always matches the minimum value of the underlying integer type.
+    pub const MIN: Self = NonMax {
+        // SAFETY: For all integers, MIN + 1 is nonzero
+        value_plus_one: unsafe { NonZero::new_unchecked(T::MIN_PLUS_ONE) },
+    };
+    /// The maximum value for this type.
+    ///
+    /// Always one less than the maximum value of the underlying type.y
+    pub const MAX: Self = NonMax {
+        // SAFETY: The maximum value is never zero.
+        value_plus_one: unsafe { NonZero::new_unchecked(T::MAX) },
+    };
+}
+
+#[cfg(feature = "bytemuck")]
+mod bytemuck_impls {
+    use super::NonMax;
+    use crate::UnsignedPrimInt;
+
+    #[cfg(feature = "bytemuck")]
+    // SAFETY: Follows from the fact we wrap a NonZero
+    unsafe impl<T: UnsignedPrimInt> bytemuck::PodInOption for NonMax<T> {}
+    #[cfg(feature = "bytemuck")]
+    // SAFETY: Follows from the fact we wrap a NonZero
+    unsafe impl<T: UnsignedPrimInt> bytemuck::ZeroableInOption for NonMax<T> {}
+    #[cfg(feature = "bytemuck")]
+    // SAFETY: Follows from the fact we wrap a NonZero
+    unsafe impl<T: UnsignedPrimInt> bytemuck::NoUninit for NonMax<T> {}
+}
+
+#[cfg(feature = "serde")]
+mod serde_impls {
+    use core::fmt::Formatter;
+    use core::marker::PhantomData;
+
+    use serde::Serializer;
+    use serde::de::{Error, Unexpected};
+
+    use super::NonMax;
+    use crate::UnsignedPrimInt;
+
+    impl<'a, T: UnsignedPrimInt> serde::Deserialize<'a> for NonMax<T> {
+        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+        where
+            D: serde::Deserializer<'a>,
+        {
+            // serde uses a visitor https://github.com/serde-rs/serde/blob/v1.0.229/serde_core/src/de/impls.rs#L92-L100
+            // this gives a better expecting() message but would be harder to implement
+            let value: T = T::deserialize(deserializer)?;
+            assert!(size_of::<T>() <= 16);
+            Self::new(value).ok_or_else(|| {
+                D::Error::invalid_value(
+                    if size_of::<T>() == 16 {
+                        Unexpected::Other(if T::SIGNED { "i128::MAX" } else { "u128::MAX" })
+                    } else if T::SIGNED {
+                        Unexpected::Signed(T::MAX.checked_cast().unwrap())
+                    } else {
+                        Unexpected::Unsigned(T::MAX.checked_cast().unwrap())
+                    },
+                    &ExpectedDesc::<T>(PhantomData),
+                )
+            })
+        }
+    }
+
+    impl<T: UnsignedPrimInt> serde::Serialize for NonMax<T> {
+        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+        where
+            S: Serializer,
+        {
+            self.get().serialize(serializer)
+        }
+    }
+
+    /// Describes what we expect, implementing [`serde::de::Expected`]
+    struct ExpectedDesc<T>(PhantomData<T>);
+    impl<T: UnsignedPrimInt> serde::de::Expected for ExpectedDesc<T> {
+        fn fmt(&self, f: &mut Formatter) -> core::fmt::Result {
+            write!(f, "a {tname} (except {tname}::MAX)", tname = T::TYPE_NAME)
+        }
+    }
+}